+++ src/tools/other.ts
@@ -749,15 +749,21 @@ export function validCountryCodeA2(code: string) {
     return helper.vars.iso._3166_1_alpha2.some(x => x == code.toUpperCase());
 }
 
-export function lazerToOldStatistics(stats: apitypes.ScoreStatistics, mode: apitypes.Ruleset): apitypes.Statistics {
+/**
+ * 
+ * @param defaultToNan - if the stat isnt found, return NaN instead of 0 
+ * @returns 
+ */
+export function lazerToOldStatistics(stats: apitypes.ScoreStatistics, mode: apitypes.Ruleset, defaultToNan?: boolean): apitypes.Statistics {
     let foo: apitypes.Statistics;
+    let dv = defaultToNan ? NaN : 0
     switch (mode) {
         case 0:
             foo = {
-                count_300: stats?.great ?? 0,
-                count_100: stats?.ok ?? 0,
-                count_50: stats?.meh ?? 0,
-                count_miss: stats?.miss ?? 0,
+                count_300: stats?.great ?? dv,
+                count_100: stats?.ok ?? dv,
+                count_50: stats?.meh ?? dv,
+                count_miss: stats?.miss ?? dv,
                 count_geki: NaN,
                 count_katu: NaN
             };
@@ -765,9 +771,9 @@ export function lazerToOldStatistics(stats: apitypes.ScoreStatistics, mode: apit
         case 1:
             foo = {
                 count_300: stats.great,
-                count_100: stats?.good ?? 0,
+                count_100: stats?.good ?? dv,
                 count_50: NaN,
-                count_miss: stats?.miss ?? 0,
+                count_miss: stats?.miss ?? dv,
                 count_geki: NaN,
                 count_katu: NaN
             };
@@ -775,21 +781,21 @@ export function lazerToOldStatistics(stats: apitypes.ScoreStatistics, mode: apit
         case 2:
             foo = {
                 count_300: stats.great, // fruits
-                count_100: stats?.ok ?? 0, // drops
-                count_50: stats?.small_tick_hit ?? 0, // droplets 
-                count_miss: stats?.miss ?? 0, //
+                count_100: stats?.ok ?? dv, // drops
+                count_50: stats?.small_tick_hit ?? dv, // droplets 
+                count_miss: stats?.miss ?? dv, //
                 count_geki: NaN,
-                count_katu: stats?.small_tick_miss ?? 0, // droplets miss
+                count_katu: stats?.small_tick_miss ?? dv, // droplets miss
             };
             break;
         case 3:
             foo = {
-                count_geki: stats?.perfect ?? 0,
+                count_geki: stats?.perfect ?? dv,
                 count_300: stats.great,
-                count_katu: stats?.good ?? 0,
-                count_100: stats?.ok ?? 0,
-                count_50: stats?.meh ?? 0,
-                count_miss: stats?.miss ?? 0,
+                count_katu: stats?.good ?? dv,
+                count_100: stats?.ok ?? dv,
+                count_50: stats?.meh ?? dv,
+                count_miss: stats?.miss ?? dv,
             };
             break;
     }